moving files for setup.py
