+++ amc2moodle/LICENSE
