+++ amc2moodle/automultiplechoice.sty.ltxml
